don't show race specific links if no race is selected
